Changed the index type in the array operator to "int" and cast inside the method to avoid warnings.
git-svn-id: https://core.forge.funambol.org/svn/core/top-level/trunk/3x/client-api/native@26826 e8e8ed6c-164c-0410-afcf-9e9a7c7d8c10

+++ src/c++/common/base/util/StringBuffer.cpp
@@ -494,9 +494,9 @@ bool  StringBuffer::operator!= (const StringBuffer& s) const
     return !(*this == s.c_str());
 }
 
-char StringBuffer::operator[]( unsigned int index ) const
+char StringBuffer::operator[]( int index ) const
 {
-    if (s && index >= 0 && index < strlen(s)) {
+    if (s && index >= 0 && (unsigned int)index < strlen(s)) {
         return s[index];
     } else {
         return (char)(-1);
